Refractor DataParser and learn to write Unit Test
