+++ DataParser.Tests/fixtures/data-directory/01-01-2022.csv
@@ -1,0 +1,3 @@
+FIPS,Admin2,Province_State,Country_Region,Last_Update,Lat,Long_,Confirmed,Deaths,Recovered,Active,Combined_Key,Incident_Rate,Case_Fatality_Ratio
+,,Henan,China,2022-01-02 04:20:52,33.882,113.61399999999999,1642,22,,,"Henan, China",1.7095262883914628,1.3398294762484775
+,,Hong Kong,China,2022-01-02 04:20:52,22.3,114.2,12667,213,,,"Hong Kong, China",168.961188146493,1.6815346964553564
