+++ DataParser.Tests/fixtures/data-directory/06-05-2020.csv
@@ -1,0 +1,3 @@
+FIPS,Admin2,Province_State,Country_Region,Last_Update,Lat,Long_,Confirmed,Deaths,Recovered,Active,Combined_Key,Incidence_Rate,Case-Fatality_Ratio
+,,Anhui,China,2020-06-06 02:33:00,31.8257,117.2264,991,6,985,0,"Anhui, China",1.6238668510457415,0.6054490413723511
+,,Hong Kong,China,2020-06-06 02:33:00,22.3,114.2,1102,4,1045,53,"Hong Kong, China",14.6992365467305,0.3629764065335753
