+++ DataParser.Tests/fixtures/format-2020-china.csv
@@ -1,0 +1,4 @@
+Province/State,Country/Region,Last Update,Confirmed,Deaths,Recovered
+Hubei,Mainland China,2020-02-21T13:03:09,62662,2144,11881
+Guangdong,Mainland China,2020-02-21T10:13:09,1333,5,690
+Hong Kong,Hong Kong,2020-02-21T03:43:02,68,2,5
