+++ covid-study.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30114.105
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataParser.Tests", "DataParser.Tests\DataParser.Tests.csproj", "{4C7A2E02-3C5F-4592-9C4F-E1E159CE2FC0}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{4C7A2E02-3C5F-4592-9C4F-E1E159CE2FC0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4C7A2E02-3C5F-4592-9C4F-E1E159CE2FC0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4C7A2E02-3C5F-4592-9C4F-E1E159CE2FC0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4C7A2E02-3C5F-4592-9C4F-E1E159CE2FC0}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
